feat: optimzize the file structure
