+++ packages/cli/doc-plugin-typedoc/src/constants.ts
@@ -1,0 +1,2 @@
+export const API_DIR = 'api';
+export const ROUTE_PREFIX = `/${API_DIR}`;
